Disable xrun for amlhalasink (#140)
Summary: Rialto was hanging because the amlhalasink sink in the server pipeline was been paused by xrun, while the rest of the pipeline remained playing. This causes issues with the termination of the pipeline as the elements states are not synchronised. The approach of disabling xrun is the same as that in wpe webkit RDKTV-16244.
Type: BugFix
Test Plan: YouTube Cert Tests
Jira: RIALTO-231