Rialto 307 change owner and group of socket (#201)
Summary: Add settings which allow specifying owner and group of the
rialto socket
Type: Feature
Test Plan: Ran unit tests. Loaded the image onto Sky0 box and changed
the appsservice code to specify a different owner and group for the
socket and the socket was created as expected. Also, a test was
conducted with the standard appservice code (not specifying an owner and
group... default empty strings) and the socket was created as expected
(with the user and primary group running the process)
Dependencies and Impacts: Extra parameters for socket owner and group
are added alongside the pre-existing socket permissions
Jira: RIALTO-307

To change the appsservice so that it specifies an owner/ group...

RDK/AppManager/RialtoServerManager/source/RialtoServerManager.cpp
@@ -59,6 +59,10 @@ RialtoServerManager::RialtoServerManager(const
std::list<std::string> &environme
     firebolt::rialto::common::ServerManagerConfig config;
     config.sessionServerEnvVars = addMandatoryEnvVars(environmentVars);
     config.numOfPreloadedServers = numOfPreloadedServers;
+
+    config.sessionManagementSocketPermissions.owner = "app-500";
+    config.sessionManagementSocketPermissions.group = "apps";
+
mServerManagerService =
rialto::servermanager::service::create(mStateObserver, config);
     if (!mServerManagerService->registerLogHandler(mLogHandler))
     {

NOTE: Marcin is now on holiday therefore the changes are being merged with a current review from Luke and an approvement of the previous version from Marcin.